+++ app/models/ResultsPageData.scala
@@ -16,8 +16,54 @@
 
 package models
 
+import connectors.MarginalReliefCalculatorConnector
 import connectors.sharedmodel.CalculatorResult
 import forms.AccountingPeriodForm
+import models.requests.DataRequest
+import pages.{ AccountingPeriodPage, AssociatedCompaniesPage, DistributionsIncludedPage, TaxableProfitPage }
+import uk.gov.hmrc.http.HeaderCarrier
+
+import scala.concurrent.{ ExecutionContext, Future }
+
+object ResultsPageData {
+  def apply(
+    marginalReliefCalculatorConnector: MarginalReliefCalculatorConnector
+  )(implicit request: DataRequest[_], hc: HeaderCarrier, ec: ExecutionContext): Future[Option[ResultsPageData]] = {
+    val maybeAccountingPeriodForm = request.userAnswers.get(AccountingPeriodPage)
+    val maybeTaxableProfit = request.userAnswers.get(TaxableProfitPage)
+    val maybeAssociatedCompanies = request.userAnswers.get(AssociatedCompaniesPage)
+    val maybeDistributionsIncludedForm = request.userAnswers.get(DistributionsIncludedPage)
+
+    (maybeAccountingPeriodForm, maybeTaxableProfit) match {
+      case (Some(accountingPeriodForm), Some(taxableProfit)) =>
+        marginalReliefCalculatorConnector
+          .calculate(
+            accountingPeriodForm.accountingPeriodStartDate,
+            accountingPeriodForm.accountingPeriodEndDate.get,
+            taxableProfit.toDouble,
+            maybeDistributionsIncludedForm.flatMap(
+              _.distributionsIncludedAmount.map(_.toDouble)
+            ),
+            maybeAssociatedCompanies.flatMap(_.associatedCompaniesCount),
+            maybeAssociatedCompanies.flatMap(_.associatedCompaniesFY1Count),
+            maybeAssociatedCompanies.flatMap(_.associatedCompaniesFY2Count)
+          )
+          .map(calculatorResult =>
+            Some(
+              ResultsPageData(
+                accountingPeriodForm,
+                taxableProfit,
+                calculatorResult,
+                maybeDistributionsIncludedForm.flatMap(_.distributionsIncludedAmount).getOrElse(0),
+                maybeAssociatedCompanies.flatMap(_.associatedCompaniesCount).getOrElse(0)
+              )
+            )
+          )
+      case _ => Future.successful(None)
+    }
+
+  }
+}
 
 case class ResultsPageData(
   accountingPeriodForm: AccountingPeriodForm,
